requirements.txt
nltk
pyphen
pdfminer.six
matplotlib